+++ hoymiles_modbus/__init__.py
@@ -1,5 +1,18 @@
 """Top-level package for hoymiles_modbus."""
 
+from .client import HoymilesDTU
+from .datatypes import InverterData, PlantData
+from .exceptions import HoymilesModbusError, InverterDataError, InvertersNotMappedError
+
 __author__ = """Mariusz Wasiluk"""
 __email__ = 'foo@bar.com'
 __version__ = '0.10.0'
+
+__all__ = [
+    'HoymilesDTU',
+    'HoymilesModbusError',
+    'InverterData',
+    'InverterDataError',
+    'InvertersNotMappedError',
+    'PlantData',
+]